% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgme-package.R
\docType{package}
\name{tgme-package}
\alias{tgme}
\alias{tgme-package}
\title{tgme: Text Message Yourself in Telegram}
\description{
This package will be familiar to fans of beepr or textme.
    The main function of this package is telegme(), which sends
    you a text message from R to your telegram bot. By adding a telegme() to a long-running
    job, you can leave the computer while it's processing, and receive 
    a text message via the instant that your job finishes. This package
    uses telegram package to send text messages, and thus requires a Telegram account..
}
\author{
\strong{Maintainer}: Anatoly Tsyplenkov \email{atsyplenkov@gmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID})

}
\keyword{internal}
