% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgme.R
\name{tgme}
\alias{tgme}
\title{Send a message to telegram bot.}
\usage{
tgme(text, bot_name = "rstudio_server", user_name = "me")
}
\arguments{
\item{text}{A message to send.}

\item{bot_name}{A name of the Telegram Bot}

\item{user_name}{A name of the user}
}
\value{
A message sent to the desired telegram bot.
}
\description{
Send a message to telegram bot.
}
\examples{

# send a custom message
tgme(text = "A new text message.")

}
