% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgme.R
\name{tgme}
\alias{tgme}
\title{Send a message to telegram bot.}
\usage{
tgme(text)
}
\arguments{
\item{text}{A message to send.}
}
\value{
A message sent to the desired telegram bot.
}
\description{
Send a message to telegram bot.
}
\examples{

# send a custom message
tgme(text = "A new text message.")

}
